fix grey rule not working,fix #6
